refactor: 폴더 구조 변경
- 관심사가 같은 컴포넌트들끼리 묶는 것으로 변경
